requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
asyncio
aiofiles==23.2.1
telethon==1.34.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9 